Moved downsample routines to separate file
